requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
groq==0.4.1
pydantic==2.5.0
langgraph==0.0.26
graphviz==0.20.1
langchain==0.0.352
langchain-community==0.0.13
chromadb==0.4.18
sentence-transformers==2.2.2
pypdf==3.17.4